This change stores a dom-if condition in a separate variable if condition is not a method call. For example: <template is="dom-if" if="[[obj.prop]]">   <div title$="[[compute(obj.prop)]]"> </template>
Before the change, the generated code is:
if(__f(obj).prop) {
  // Typescript doesn't infer that __f(obj).prop is defined her:
  ...compute(__f(obj).prop)
}

After the change, the __f(obj).prop is store in a constant and
typescript infers type inside if correctly.

We don't store the result of a method call in variable: often a method
in an condtion is a type guard and it can't be extracted to a varible.